Refactor domain models and add theme modes including Telegram dark support
