+++ src/app/auth/auth.component.scss
@@ -11,8 +11,8 @@
 .login-form {
   width: min(440px, 100%);
   margin: 9vh auto 0;
-  background: var(--tg-theme-secondary-bg-color, #fff);
-  color: var(--tg-theme-text-color, #111827);
+  background: var(--app-surface);
+  color: var(--app-text);
   border: 1px solid #d1d5db;
   border-radius: 16px;
   padding: 1.25rem;
@@ -45,7 +45,7 @@ h3 {
 
 label {
   font-size: 0.9rem;
-  color: var(--tg-theme-hint-color, #374151);
+  color: var(--app-muted-text);
 }
 
 input {
@@ -53,8 +53,8 @@ input {
   border-radius: 10px;
   padding: 0.65rem 0.75rem;
   font-size: 0.95rem;
-  color: var(--tg-theme-text-color, #111827);
-  background: var(--tg-theme-bg-color, #fff);
+  color: var(--app-text);
+  background: var(--app-surface);
 }
 
 .actions {
