+++ src/app/game/game.component.scss
@@ -41,7 +41,7 @@
 details {
   border: 1px solid #e5e7eb;
   border-radius: 12px;
-  background: var(--tg-theme-secondary-bg-color, #fff);
+  background: var(--app-surface);
   padding: 0.75rem;
   margin-bottom: 0.85rem;
 }
