+++ src/app/header/header.component.scss
@@ -1,8 +1,8 @@
 @import "../../styles";
 
 .active-game-banner {
-  background-color: #205d35;
-  color: #f6fff8;
+  background-color: var(--app-header-bg);
+  color: var(--app-header-text);
   font-size: 0.9rem;
   padding: 0.5rem 1rem;
 }
@@ -11,13 +11,13 @@
   position: sticky;
   top: 0;
   z-index: 15;
-  background: rgba(19, 92, 43, 0.95);
+  background: color-mix(in srgb, var(--app-header-bg) 90%, transparent);
   backdrop-filter: blur(8px);
-  color: #fff;
+  color: var(--app-header-text);
   min-height: 64px;
   padding: 0.75rem 1rem;
   display: grid;
-  grid-template-columns: auto 1fr auto;
+  grid-template-columns: auto 1fr auto auto;
   gap: 1rem;
   align-items: center;
 }
@@ -37,7 +37,7 @@
 
 .nav-link {
   text-decoration: none;
-  color: #e8fff0;
+  color: inherit;
   padding: 0.45rem 0.75rem;
   border-radius: 999px;
 }
@@ -47,11 +47,25 @@
   background: rgba(255, 255, 255, 0.18);
 }
 
+.theme-picker,
 .user-section,
 .user-info {
   display: flex;
   align-items: center;
   gap: 0.75rem;
+}
+
+.theme-label {
+  font-size: 0.9rem;
+  opacity: 0.92;
+}
+
+.theme-select {
+  border-radius: 10px;
+  border: 1px solid var(--app-border);
+  background: var(--app-surface);
+  color: var(--app-text);
+  padding: 0.35rem 0.5rem;
 }
 
 .avatar-link {
@@ -73,7 +87,7 @@
 }
 
 .ghost-button {
-  color: #fff;
+  color: var(--app-header-text);
   background: rgba(255, 255, 255, 0.16);
 }
 
@@ -86,7 +100,7 @@
   display: none;
   justify-self: start;
   background: transparent;
-  color: #fff;
+  color: inherit;
   border: none;
 }
 
@@ -101,7 +115,8 @@
   }
 
   .desktop-nav,
-  .user-section {
+  .user-section,
+  .theme-picker {
     display: none;
   }
 
@@ -133,14 +148,19 @@
     left: 0;
     height: 100dvh;
     width: min(86vw, 320px);
-    background: var(--tg-theme-secondary-bg-color, #fff);
-    color: var(--tg-theme-text-color, #111827);
+    background: var(--app-surface);
+    color: var(--app-text);
     z-index: 19;
     transform: translateX(-100%);
     transition: transform 0.25s ease;
     flex-direction: column;
     padding: 1rem;
     gap: 0.75rem;
+  }
+
+  .mobile-theme-picker {
+    display: flex;
+    justify-content: space-between;
   }
 
   .mobile-menu.open {
@@ -151,14 +171,15 @@
     align-self: flex-end;
     border: none;
     background: transparent;
+    color: inherit;
   }
 
   .mobile-link {
     text-decoration: none;
-    color: var(--tg-theme-text-color, #1f2937);
+    color: var(--app-text);
     font-weight: 600;
     padding: 0.5rem 0;
-    border-bottom: 1px solid #e5e7eb;
+    border-bottom: 1px solid var(--app-border);
   }
 
   .mobile-auth {
@@ -170,7 +191,7 @@
 
   .mobile-username {
     margin: 0;
-    color: var(--tg-theme-text-color, #374151);
+    color: var(--app-muted-text);
   }
 
   .mobile-auth .ghost-button {
