+++ src/app/hint.part/hint.part.component.ts
@@ -1,5 +1,5 @@
 import {Component, Input} from '@angular/core';
-import {HintPart, HintType} from "../game/game.service";
+import {HintPart, HintType} from "../domain/game.models";
 
 
 @Component({
