+++ src/app/home/home.component.scss
@@ -11,6 +11,6 @@ h1 {
 p {
   margin: 0 auto;
   max-width: 60ch;
-  color: #4b5563;
+  color: var(--app-muted-text);
   line-height: 1.5;
 }
