+++ src/app/profile/profile.component.scss
@@ -3,9 +3,9 @@
 .profile-page {
   max-width: 560px;
   margin: 0 auto;
-  background: var(--tg-theme-bg-color, #ffffff);
-  color: var(--tg-theme-text-color, #1f2937);
-  border: 1px solid var(--tg-theme-hint-color, #dbe4dd);
+  background: var(--app-surface);
+  color: var(--app-text);
+  border: 1px solid var(--app-border);
   border-radius: 16px;
   padding: 1.25rem;
 }
@@ -32,12 +32,12 @@ label {
 }
 
 input {
-  border: 1px solid var(--tg-theme-hint-color, #cbd5e1);
+  border: 1px solid var(--app-border);
   border-radius: 10px;
   padding: 0.65rem 0.75rem;
   font-size: 0.95rem;
-  color: var(--tg-theme-text-color, #1f2937);
-  background: var(--tg-theme-secondary-bg-color, #ffffff);
+  color: var(--app-text);
+  background: var(--app-surface);
 }
 
 .primary {
